+++ connect/pages/Welcome.js
